Recompose the driver's test files
We now have a cleaner separation between topics, services, and actions.
